// schemas/public/UserProfile.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { z } from 'zod';

/** Identifier type for public.user_profile */
export type UserProfileId = string & { __brand: 'UserProfileId' };

/** Represents the table public.user_profile */
export default interface UserProfile {
  id: UserProfileId;

  email: string;

  password: string;

  is_anonymous: boolean;

  is_admin: boolean;

  is_active: boolean;

  is_subscribed: boolean;

  x_created: Date;

  first_name: string;

  last_name: string;
}

/** Represents the initializer for the table public.user_profile */
export interface UserProfileInitializer {
  /** Default value: uuid_generate_v4() */
  id?: UserProfileId;

  email: string;

  password: string;

  is_anonymous: boolean;

  is_admin: boolean;

  is_active: boolean;

  is_subscribed: boolean;

  /** Default value: CURRENT_TIMESTAMP */
  x_created?: Date;

  first_name: string;

  last_name: string;
}

/** Represents the mutator for the table public.user_profile */
export interface UserProfileMutator {
  id?: UserProfileId;

  email?: string;

  password?: string;

  is_anonymous?: boolean;

  is_admin?: boolean;

  is_active?: boolean;

  is_subscribed?: boolean;

  x_created?: Date;

  first_name?: string;

  last_name?: string;
}

export const userProfileId = z.string() as unknown as z.Schema<UserProfileId>;

export const userProfile = z.object({
  id: userProfileId,
  email: z.string(),
  password: z.string(),
  is_anonymous: z.boolean(),
  is_admin: z.boolean(),
  is_active: z.boolean(),
  is_subscribed: z.boolean(),
  x_created: z.date(),
  first_name: z.string(),
  last_name: z.string(),
}) as unknown as z.Schema<UserProfile>;

export const userProfileInitializer = z.object({
  id: userProfileId.optional(),
  email: z.string(),
  password: z.string(),
  is_anonymous: z.boolean(),
  is_admin: z.boolean(),
  is_active: z.boolean(),
  is_subscribed: z.boolean(),
  x_created: z.date().optional(),
  first_name: z.string(),
  last_name: z.string(),
}) as unknown as z.Schema<UserProfileInitializer>;

export const userProfileMutator = z.object({
  id: userProfileId.optional(),
  email: z.string().optional(),
  password: z.string().optional(),
  is_anonymous: z.boolean().optional(),
  is_admin: z.boolean().optional(),
  is_active: z.boolean().optional(),
  is_subscribed: z.boolean().optional(),
  x_created: z.date().optional(),
  first_name: z.string().optional(),
  last_name: z.string().optional(),
}) as unknown as z.Schema<UserProfileMutator>;