// schemas/public/BlogPost.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { blogPostStatusId, type BlogPostStatusId } from './BlogPostStatus';
import { z } from 'zod';

/** Identifier type for public.blog_post */
export type BlogPostId = string & { __brand: 'BlogPostId' };

/** Represents the table public.blog_post */
export default interface BlogPost {
  id: BlogPostId;

  title: string;

  seo_name: string;

  seo_description: string;

  seo_keywords: string;

  status: BlogPostStatusId;

  x_created: Date;

  content: string | null;
}

/** Represents the initializer for the table public.blog_post */
export interface BlogPostInitializer {
  /** Default value: uuid_generate_v4() */
  id?: BlogPostId;

  title: string;

  seo_name: string;

  seo_description: string;

  seo_keywords: string;

  status: BlogPostStatusId;

  /** Default value: CURRENT_TIMESTAMP */
  x_created?: Date;

  content?: string | null;
}

/** Represents the mutator for the table public.blog_post */
export interface BlogPostMutator {
  id?: BlogPostId;

  title?: string;

  seo_name?: string;

  seo_description?: string;

  seo_keywords?: string;

  status?: BlogPostStatusId;

  x_created?: Date;

  content?: string | null;
}

export const blogPostId = z.string() as unknown as z.Schema<BlogPostId>;

export const blogPost = z.object({
  id: blogPostId,
  title: z.string(),
  seo_name: z.string(),
  seo_description: z.string(),
  seo_keywords: z.string(),
  status: blogPostStatusId,
  x_created: z.date(),
  content: z.string().nullable(),
}) as unknown as z.Schema<BlogPost>;

export const blogPostInitializer = z.object({
  id: blogPostId.optional(),
  title: z.string(),
  seo_name: z.string(),
  seo_description: z.string(),
  seo_keywords: z.string(),
  status: blogPostStatusId,
  x_created: z.date().optional(),
  content: z.string().optional().nullable(),
}) as unknown as z.Schema<BlogPostInitializer>;

export const blogPostMutator = z.object({
  id: blogPostId.optional(),
  title: z.string().optional(),
  seo_name: z.string().optional(),
  seo_description: z.string().optional(),
  seo_keywords: z.string().optional(),
  status: blogPostStatusId.optional(),
  x_created: z.date().optional(),
  content: z.string().optional().nullable(),
}) as unknown as z.Schema<BlogPostMutator>;