// schemas/public/Image.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { z } from 'zod';

/** Identifier type for public.image */
export type ImageId = string & { __brand: 'ImageId' };

/** Represents the table public.image */
export default interface Image {
  id: ImageId;

  data: string;

  name: string;

  alt: string;

  metadata: string | null;

  x_created: Date;
}

/** Represents the initializer for the table public.image */
export interface ImageInitializer {
  /** Default value: uuid_generate_v4() */
  id?: ImageId;

  data: string;

  name: string;

  alt: string;

  metadata?: string | null;

  /** Default value: now() */
  x_created?: Date;
}

/** Represents the mutator for the table public.image */
export interface ImageMutator {
  id?: ImageId;

  data?: string;

  name?: string;

  alt?: string;

  metadata?: string | null;

  x_created?: Date;
}

export const imageId = z.string() as unknown as z.Schema<ImageId>;

export const image = z.object({
  id: imageId,
  data: z.string(),
  name: z.string(),
  alt: z.string(),
  metadata: z.string().nullable(),
  x_created: z.date(),
}) as unknown as z.Schema<Image>;

export const imageInitializer = z.object({
  id: imageId.optional(),
  data: z.string(),
  name: z.string(),
  alt: z.string(),
  metadata: z.string().optional().nullable(),
  x_created: z.date().optional(),
}) as unknown as z.Schema<ImageInitializer>;

export const imageMutator = z.object({
  id: imageId.optional(),
  data: z.string().optional(),
  name: z.string().optional(),
  alt: z.string().optional(),
  metadata: z.string().optional().nullable(),
  x_created: z.date().optional(),
}) as unknown as z.Schema<ImageMutator>;